+++ shardingsphere-governance/shardingsphere-governance-core/src/test/java/org/apache/shardingsphere/governance/core/registry/watcher/impl/DataSourceStateChangedWatcherTest.java
@@ -30,7 +30,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
-public final class DataSourceStateChangedListenerTest {
+public final class DataSourceStateChangedWatcherTest {
     
     @Test
     public void assertCreateEvent() {
