+++ shardingsphere-governance/shardingsphere-governance-core/src/test/java/org/apache/shardingsphere/governance/core/registry/watcher/impl/LockChangedWatcherTest.java
@@ -17,18 +17,20 @@
 
 package org.apache.shardingsphere.governance.core.registry.watcher.impl;
 
-import lombok.SneakyThrows;
+import org.apache.shardingsphere.governance.core.registry.watcher.event.GovernanceEvent;
+import org.apache.shardingsphere.governance.repository.api.listener.DataChangedEvent;
+import org.apache.shardingsphere.governance.repository.api.listener.DataChangedEvent.Type;
+import org.junit.Test;
 
-import java.io.IOException;
-import java.net.URISyntaxException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.stream.Collectors;
+import java.util.Optional;
 
-public class GovernanceListenerTest {
+import static org.junit.Assert.assertFalse;
+
+public final class LockChangedWatcherTest {
     
-    @SneakyThrows({IOException.class, URISyntaxException.class})
-    protected String readYAML(final String yamlFile) {
-        return Files.readAllLines(Paths.get(ClassLoader.getSystemResource(yamlFile).toURI())).stream().map(each -> each + System.lineSeparator()).collect(Collectors.joining());
+    @Test
+    public void assertCreateEventWithInvalidPath() {
+        Optional<GovernanceEvent> actual = new LockChangedWatcher().createGovernanceEvent(new DataChangedEvent("/lock/glock", "", Type.ADDED));
+        assertFalse(actual.isPresent());
     }
 }
